# lib/api-zod/src/generated/types/leadDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VisaCRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeadDocument } from './leadDocument';
import type { LeadNote } from './leadNote';
import type { LeadPayment } from './leadPayment';
import type { LeadStatusHistory } from './leadStatusHistory';

export interface LeadDetail {
  id: string;
  pax_name: string;
  phone: string;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  service_id?: string | null;
  service_name: string;
  /** @nullable */
  sub_service?: string | null;
  /** @nullable */
  destination_country?: string | null;
  source: string;
  status: string;
  /** @nullable */
  assigned_to?: string | null;
  /** @nullable */
  agent_name?: string | null;
  base_fee: number;
  gst_amount: number;
  total_amount: number;
  amount_paid: number;
  payment_method: string;
  /** @nullable */
  notes?: string | null;
  created_at: string;
  updated_at: string;
  lead_notes: LeadNote[];
  lead_payments: LeadPayment[];
  lead_documents: LeadDocument[];
  lead_status_history: LeadStatusHistory[];
}