# lib/api-zod/src/generated/types/leadStatusHistory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VisaCRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeadStatusHistory {
  id: string;
  lead_id: string;
  status: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  changed_by?: string | null;
  /** @nullable */
  changed_by_name?: string | null;
  created_at: string;
}