# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VisaCRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeadUpdate {
  pax_name?: string;
  phone?: string;
  whatsapp?: string;
  email?: string;
  service_id?: string;
  service_name?: string;
  sub_service?: string;
  destination_country?: string;
  source?: string;
  status?: string;
  assigned_to?: string;
  agent_name?: string;
  base_fee?: number;
  amount_paid?: number;
  payment_method?: string;
  notes?: string;
}