# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VisaCRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  full_name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  role_id?: string | null;
  role_name: string;
  status: string;
  avatar_color: string;
  created_at: string;
}