# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VisaCRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Role {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  is_preset: boolean;
  permissions: string[];
  created_at: string;
}